moved custom schemas from client to validators, cleaned up loose ends
