+++ apps/nextjs/src/app/(authenticated)/messages/_components/create-message.tsx
@@ -1,5 +1,4 @@
 "use client";
-"use client";
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
