+++ apps/nextjs/src/app/(authenticated)/messages/layout.tsx
@@ -22,4 +22,3 @@ export default function MessagesLayout({
     </>
   );
 }
-
