+++ apps/nextjs/src/app/(authenticated)/patient/layout.tsx
@@ -14,4 +14,3 @@ export default function PatientLayout({
     </div>
   );
 }
-
