+++ apps/nextjs/src/app/(authenticated)/patient/page.tsx
@@ -1,7 +1,5 @@
 const PatientPage = () => {
-  return (
-    <div>PatientPage</div>
-  )
-}
+  return <div>PatientPage</div>;
+};
 
-export default PatientPage
+export default PatientPage;
