+++ apps/nextjs/src/app/(landing)/_components/posts.tsx
@@ -129,14 +129,16 @@ export function PostCardSkeleton(props: { pulse?: boolean }) {
     <div className="flex flex-row rounded-lg bg-white/10 p-4 transition-all hover:scale-[101%]">
       <div className="flex-grow">
         <h2
-          className={`w-1/4 rounded bg-pink-400 text-2xl font-bold ${pulse && "animate-pulse"
-            }`}
+          className={`w-1/4 rounded bg-pink-400 text-2xl font-bold ${
+            pulse && "animate-pulse"
+          }`}
         >
           &nbsp;
         </h2>
         <p
-          className={`mt-2 w-1/3 rounded bg-current text-sm ${pulse && "animate-pulse"
-            }`}
+          className={`mt-2 w-1/3 rounded bg-current text-sm ${
+            pulse && "animate-pulse"
+          }`}
         >
           &nbsp;
         </p>
