+++ packages/api/src/router/canvas.ts
@@ -1,4 +1,3 @@
-import { TRPCClientError } from "@trpc/client";
 import { TRPCError } from "@trpc/server";
 import { z } from "zod";
